回退 'Pull Request !90173 : [BUGFIX]Fix std::accumulate overflow'

+++ mindspore/ops/infer/dense_to_sparse_set_operation.cc
@@ -78,8 +78,7 @@ abstract::TupleShapePtr DenseToSparseSetOperationInferShape(const PrimitivePtr &
 
   std::string set_operation_str = GetValue<std::string>(primitive->GetAttr("set_operation"));
   (void)std::transform(set_operation_str.begin(), set_operation_str.end(), set_operation_str.begin(), ::tolower);
-  int64_t x1_size =
-    std::accumulate(x1_shape.begin(), x1_shape.end(), static_cast<int64_t>(1), std::multiplies<int64_t>());
+  int64_t x1_size = std::accumulate(x1_shape.begin(), x1_shape.end(), 1, std::multiplies<int64_t>());
   int64_t x2_size = SizeToLong(x2_values_shape[LongToSize(0)]);
 
   int64_t y_size_max = 0;
