+++ mindspore/ops/infer/ops_func_impl/unique_consecutive.cc
@@ -37,8 +37,7 @@ ShapeArray UniqueConsecutiveFuncImpl::InferShape(const PrimitivePtr &primitive,
   }
 
   if (input_infos[kInputIndex3]->IsNone()) {
-    auto y_max_shape =
-      std::accumulate(shape_x.begin(), shape_x.end(), static_cast<int64_t>(1), std::multiplies<int64_t>());
+    auto y_max_shape = std::accumulate(shape_x.begin(), shape_x.end(), 1, std::multiplies<int64_t>());
     auto out_shape = x->IsDynamic() ? ShapeVector{abstract::Shape::kShapeDimAny} : ShapeVector{y_max_shape};
     auto inverse_indices_shape = (return_inverse.has_value() && return_inverse.value()) ? shape_x : empty_shape;
     auto counts_shape = (return_counts.has_value() && return_counts.value()) ? out_shape : empty_shape;
