+++ mindspore/ops/kernel/ascend/aicpu/aicpu_ops/cpu_kernel/ms_kernel/fft_ortho.cc
@@ -42,8 +42,7 @@ const char *kFFTOrtho = "FFTOrtho";
 template <typename T>
 bool Orthogonalize(T *input, T *output, const std::vector<int64_t> &input_shape, const std::vector<int64_t> dims,
                    bool forward) {
-  int64_t input_nums =
-    std::accumulate(input_shape.begin(), input_shape.end(), static_cast<int64_t>(1), std::multiplies<int64_t>());
+  int64_t input_nums = std::accumulate(input_shape.begin(), input_shape.end(), 1, std::multiplies<int64_t>());
   std::set<int64_t> ortho_dims(dims.begin(), dims.end());
   T normal_factor{2};
   T head_factor{4};
@@ -52,8 +51,7 @@ bool Orthogonalize(T *input, T *output, const std::vector<int64_t> &input_shape,
   std::vector<int64_t> offsets(input_shape.size(), 0);
   for (size_t j = 0; j < input_shape.size(); j++) {
     int64_t pos = static_cast<int64_t>(j);
-    offsets[j] =
-      std::accumulate(input_shape.begin() + pos + 1, input_shape.end(), static_cast<int64_t>(1), std::multiplies<>());
+    offsets[j] = std::accumulate(input_shape.begin() + pos + 1, input_shape.end(), 1, std::multiplies<>());
   }
 
   for (int64_t i = 0; i < input_nums; ++i) {
