+++ mindspore/ops/kernel/ascend/aicpu/aicpu_ops/cpu_kernel/ms_kernel/irfft_double.cc
@@ -43,8 +43,7 @@ constexpr int kOnsideDivisor = 2;
 
 template <typename T>
 bool PartialDouble(T *input, T *output, const std::vector<int64_t> &input_shape, int64_t n, int64_t dim) {
-  int64_t input_nums =
-    std::accumulate(input_shape.begin(), input_shape.end(), static_cast<int64_t>(1), std::multiplies<int64_t>());
+  int64_t input_nums = std::accumulate(input_shape.begin(), input_shape.end(), 1, std::multiplies<int64_t>());
   int64_t start_pos{1};
   int64_t end_pos = start_pos + n - (n / kOnsideDivisor + 1);
 
@@ -52,8 +51,7 @@ bool PartialDouble(T *input, T *output, const std::vector<int64_t> &input_shape,
   std::vector<int64_t> offsets(input_shape.size(), 0);
   for (size_t j = 0; j < input_shape.size(); j++) {
     int64_t pos = static_cast<int64_t>(j);
-    offsets[j] =
-      std::accumulate(input_shape.begin() + pos + 1, input_shape.end(), static_cast<int64_t>(1), std::multiplies<>());
+    offsets[j] = std::accumulate(input_shape.begin() + pos + 1, input_shape.end(), 1, std::multiplies<>());
   }
 
   for (int64_t i = 0; i < input_nums; ++i) {
